+++ dinau/__init__.py
@@ -22,4 +22,4 @@ __all__ = [
     "AirQualityIndex",
 ]
 
-__version__ = "0.1.2"
+__version__ = "1.0.0"